chore(v0.2.0): ruff format + autofix sweep across v0.2.0 changes
Ran 'uv run ruff format' and 'uv run ruff check --fix' over the
feature/v0.2.0 working tree. Pure formatting + autofix; no behavior
changes. Also added a module docstring to scripts/__init__.py to
satisfy D104.

Verified:
- pytest: 145 passed
- ruff format --check: clean
- ruff check: clean
- mypy (src/ per pyproject config): clean

+++ tests/conftest.py
@@ -18,8 +18,8 @@ def _install_mlx_memory_caps() -> None:
     except ImportError:  # pragma: no cover - MLX always present on Apple Silicon
         return
     try:
-        mx.set_wired_limit(int(20 * 1024**3))
-        mx.set_memory_limit(int(22 * 1024**3))
+        mx.set_wired_limit(20 * 1024**3)
+        mx.set_memory_limit(22 * 1024**3)
     except Exception:  # pragma: no cover - older MLX / non-Metal env
         return
